changes {assertive}	R Documentation

Important changes to assertive

Description

Changes since 0.3-0:

Virtuality

The assertive package is now a virtual package: that is, it no longer contains its own functions, but instead rexports them from lower-level packages.

For interactive use, you can carry on using assertive as before.

For programmatic use, you can have more fine-grained control over what gets loaded by using the lower-level packages.

assertive.base contains the core functionality. assertive.properties contains checks on properties of variables. assertive.types contains checks on types of variables. assertive.numbers contains checks for numbers. assertive.strings contains checks for strings. assertive.datetimes contains checks for dates and times. assertive.files contains checks for files and connections. assertive.sets contains checks for sets. assertive.matrices contains checks for matrices. assertive.models contains checks for models. assertive.data contains checks for complex data types. assertive.data.uk contains checks for UK-specific complex data types. assertive.data.us contains checks for US-specific complex data types. assertive.reflection contains checks on the state of R. assertive.code contains checks for code.

Translations

The infrastructure for errors and warnings in multiple langauges is in place, and translations are planned for future versions. If you want to be a translator, email me at richierocks@gmail.com.

[Package assertive version 0.3-5 Index]