is_if_condition {assertive}	R Documentation

Is suitable to be used as an if condition See is_if_condition.

Description

Is suitable to be used as an if condition

See is_if_condition.

[Package assertive version 0.3-5 Index]